requirements.txt
streamlit-nightly
openai
llama-index
nltk
